chore(i18n) small refactor
