+++ src/i18n/i18n.ts
@@ -74,7 +74,6 @@ i18n.use(LanguageDetector).use(initReactI18next); // passes i18n down to react-i
 if (!isDev) {
   i18n.use(HttpApi);
 }
-console.log("dd", defaultBundle);
 
 const packResources = (bundleContent: any) => ({
   [DEFAULT_LOCALE]: { translation: bundleContent },
@@ -109,12 +108,15 @@ export default i18n;
 // HMR for english locale - new keys will be firstly added to en.json
 // then reuploaded to crowdin and translated
 if (isDev && (module as any).hot) {
-  (module as any).hot.accept(`./locales/en.json`, (stuff: any) => {
-    const newBundle = require("./locales/en.json");
-    console.log("aaa", stuff, newBundle);
-    i18n.removeResourceBundle(DEFAULT_LOCALE, "translation");
-    i18n.addResourceBundle(DEFAULT_LOCALE, "translation", newBundle);
-    i18n.reloadResources(["en"]).finally();
-    i18n.changeLanguage(DEFAULT_LOCALE).finally();
-  });
+  (module as any).hot.accept(
+    `./locales/${DEFAULT_LOCALE}.json`,
+    (stuff: any) => {
+      const newBundle = require(`./locales/${DEFAULT_LOCALE}.json`);
+      console.log("aaa", stuff, newBundle);
+      i18n.removeResourceBundle(DEFAULT_LOCALE, "translation");
+      i18n.addResourceBundle(DEFAULT_LOCALE, "translation", newBundle);
+      i18n.reloadResources([DEFAULT_LOCALE]).finally();
+      i18n.changeLanguage(DEFAULT_LOCALE).finally();
+    }
+  );
 }
